refactor(bbr): move test helpers into plugins/test
