+++ pkg/bbr/handlers/server_test.go
@@ -29,7 +29,7 @@ import (
 
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/bbr/framework"
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/bbr/plugins"
-	bbrtest "sigs.k8s.io/gateway-api-inference-extension/pkg/bbr/test"
+	"sigs.k8s.io/gateway-api-inference-extension/pkg/bbr/plugins/test"
 	envoytest "sigs.k8s.io/gateway-api-inference-extension/pkg/common/envoy/test"
 	logutil "sigs.k8s.io/gateway-api-inference-extension/pkg/common/observability/logging"
 )
@@ -130,7 +130,7 @@ func TestHandleRequestBodyStreaming(t *testing.T) {
 			},
 		},
 	}
-	baseModelToHeaderPlugin, err := bbrtest.NewTestBaseModelPlugin()
+	baseModelToHeaderPlugin, err := test.NewTestBaseModelPlugin()
 	if err != nil {
 		t.Fatalf("failed to create base model plugin: %v", err)
 	}
