+++ pkg/bbr/plugins/test/base_model_plugin.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package bbrtest
+package test
 
 import (
 	"fmt"
